python_service/src/dash/requirements.txt
dash==2.17.1
pandas==2.2.2
plotly==5.22.0
sqlalchemy==2.0.31
python-dotenv==1.0.1
dash-bootstrap-components==1.6.0
numpy==1.26.4
